refactor: feedLayout.tsx에서 React Hooks 분리
